🎨Refactor:유저정보폼 구조 변경
기존에 회원가입과 정보수정에서 공유 변수를 사용했으나
비효율적이라 생각하여 각자 따로 변수를 사용하도록 수정
