+++ src/components/Desktop/MyPage/MyPage.jsx
@@ -2,23 +2,72 @@ import React, { useState } from "react";
 import UserInfoForm from "../commons/UserForm/UserInfoForm";
 import { Box, Button } from "@mui/material";
 import { useNavigate } from "react-router-dom";
-import useUserValidation from "../commons/UserForm/useUserValidation";
 
 export default function MyPage() {
   const [editState, setEditState] = useState(false);
   const [isMyPage] = useState(true);
   const [pwCheck, setPwCheck] = useState("");
   const navigate = useNavigate();
+
+  //기본 정보 통신으로 받아와야함
+  const [userData, setUserData] = useState({
+    id: "홍길동",
+    pw: "",
+    pwCheck: "",
+    userName: "",
+    account: "",
+    phoneNumber: "",
+  });
+
+  const [errors, setErrors] = useState({
+    id: false,
+    pw: false,
+    pwCheck: false,
+    userName: false,
+    account: false,
+    phoneNumber: false,
+  });
+  const validate = () => {
+    let newErrors = {
+      id: userData.id === "",
+      pw: userData.pw === "",
+      pwCheck: userData.pwCheck !== userData.pw,
+      userName: userData.userName === "",
+      account: userData.account === "",
+      phoneNumber: userData.phoneNumber.length !== 11,
+    };
+
+    setErrors(newErrors);
+    // 에러가 하나라도 있는지 검사
+    return !Object.values(newErrors).some((error) => error);
+  };
+
+  const handleSubmit = (e) => {
+    e.preventDefault();
+    if (validate()) {
+      alert("수정되었습니다.");
+      console.log(userData);
+      navigate("/OrderManage");
+    } else {
+      alert("정보를 확인해주세요.");
+    }
+  };
+
+  const handleChange = (e) => {
+    const { id, value } = e.target;
+    setUserData((prevData) => ({ ...prevData, [id]: value }));
+    setErrors((prevErrors) => ({ ...prevErrors, [id]: false }));
+  };
+
   const handleBack = () => {
     if (confirm("정보수정을 취소하시겠습니까?")) {
-      navigate(-1);
+      navigate("/OrderManage");
     }
   };
   //수정모드에서 수정버튼 클릭시
   const Edit = () => {
     setEditState(!editState);
   };
-  const { handleChange, handleSubmit, userData, errors } = useUserValidation();
   return (
     <div>
       <Button
